doc strings and cleanup
