+++ packages/core/src/protocol/CounterPartyData.ts
@@ -20,6 +20,12 @@ export type CounterPartyDataOptions = z.infer<
   typeof CounterPartyDataOptionsSchema
 >;
 
+/**
+ * this function serializes CounterPartyDataOptions into bytes.  The options are sorted
+ * to match other platforms
+ * @param options - incoming CounterPartyDataOptions
+ * @returns Uint8Array representing byte encoded string of options
+ */
 export function counterPartyDataOptionsToBytes(options: CounterPartyDataOptions): Uint8Array {
   let formatArray = new Array<string>();
   Object.keys(options).sort().forEach((key) => {
@@ -27,7 +33,6 @@ export function counterPartyDataOptionsToBytes(options: CounterPartyDataOptions)
       formatArray.push(`${key}:${options[k].mandatory ? "1" : "0"}`);
   })
   let formatStr = formatArray.join(",");
-  console.log(formatStr);
   return new TextEncoder().encode(formatStr);
 }
