+++ packages/core/src/serializerUtils.ts
@@ -1,26 +1,36 @@
-export function serializeNumber(value: number): Uint8Array {
+/**
+ * converts a number into a byte representation
+ * @param value - number or bigint to serialize
+ * @returns 
+ */
+export function serializeNumber(value: number | bigint): Uint8Array {
     let result = new Uint8Array();
     if (Number.isInteger(value)) {
         if (value >= -128 && value <= 127) { // uint 8
             const buffer = new ArrayBuffer(1);
             const view = new DataView(buffer);
-            view.setInt8(0, value);
+            view.setInt8(0, value as number);
             result = new Uint8Array(buffer);
         } else if (value >= -32768 && value <= 32767) { // uint 16
             const buffer = new ArrayBuffer(2);
             const view = new DataView(buffer);
-            view.setInt16(0, value);
+            view.setInt16(0, value as number);
             result = new Uint8Array(buffer);
-    } else if (value >= -2147483648 && value <= 2147483647) { // unint 32
+        } else if (value >= -2147483648 && value <= 2147483647) { // unint 32
             const buffer = new ArrayBuffer(4);
             const view = new DataView(buffer);
-            view.setInt32(0, value);
+            view.setInt32(0, value as number);
+            result = new Uint8Array(buffer);
+        } else {
+            const buffer = new ArrayBuffer(8);
+            const view = new DataView(buffer);
+            view.setBigUint64(0, value as bigint);
             result = new Uint8Array(buffer);
         }
     } else {
         const buffer = new ArrayBuffer(4);
         const view = new DataView(buffer);
-        view.setFloat32(0, value);
+        view.setFloat32(0, value as number);
         result = new Uint8Array(buffer);
     }
     return result;
@@ -34,7 +44,13 @@ export function serializeBoolean(value: boolean): Uint8Array {
     return new Uint8Array([value ? 1 : 0]);
 }
 
-export function deserializeNumber(value: Uint8Array): number {
+/**
+ * there isn't enough info to infer int / float from a serialized number, so we recommend providing an explicit float
+ * deserialize function for values intended to be integers or bigints.
+ * @param value - Uint8Array representing a number serialized as bytes
+ * @returns integer or bigint
+ */
+export function deserializeNumber(value: Uint8Array): number | bigint {
     let result;
     const length = value.length;
     const view = new DataView(value.buffer);
@@ -51,14 +67,41 @@ export function deserializeNumber(value: Uint8Array): number {
             result = view.getInt32(0);
             break;
         }
-        // case 8 : { // 64 bit
-        //     result = view.getBigInt64(0);
-        //     break;
-        // }
+        case 8 : { // 64 bit
+            result = view.getBigInt64(0);
+            break;
+        }
         default: {
             result = view.getInt8(0);
             break;
         } break;
+    }
+    return result;
+}
+
+/**
+ * there isn't enough info to infer int / float from a serialized number, so we recommend providing an explicit float
+ * deserialize function for values intended to be floats or doubles.
+ * @param value - Uint8Array representing byte serialization of a float
+ * @returns number either 8 byte or 4 byte float. if lenght of input value is not 4 or 8, return the first byte 
+ */
+export function deserializeFloat(value: Uint8Array): number {
+    let result;
+    const length = value.length;
+    const view = new DataView(value.buffer);
+    switch (length) {
+        case 4: {
+            result = view.getFloat32(0);
+            break;
+        }
+        case 8: {
+            result = view.getFloat64(0);
+            break;
+        }
+        default: {
+            result = view.getUint8(0);
+            break;
+        }
     }
     return result;
 }
